pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pybeli"
version = "0.1.0"
description = "A CLI tool for analyzing stock and crypto tickers using technical analysis indicators."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Beerware" }
authors = [
    { name = "Ernest Duckworth IV", email = "rionduckworth@gmail.com" },
]
keywords = ["cli", "stocks", "crypto", "technical-analysis", "trading"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "pyaml>=26.2.1",
    "pydantic>=2.12.5",
    "types-pyyaml>=6.0.12.20260408",
    "yfinance>=1.2.1",
]

[project.urls]
Homepage = "https://github.com/ernman37/pibeli"
Repository = "https://github.com/ernman37/pibeli"
Issues = "https://github.com/ernman37/pibeli/issues"

[project.scripts]
pybeli = "pybeli.cli.main:main"

[dependency-groups]
dev = [
    "mypy>=1.0",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "pre-commit>=3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pybeli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=src/pybeli --cov-report=term-missing --cov-fail-under=80"

[tool.mypy]
strict = true
python_version = "3.12"
files = ["src"]

[[tool.mypy.overrides]]
module = "yfinance.*"
ignore_missing_imports = true
